+++ components/restreamer/src/lib.rs
@@ -40,6 +40,7 @@ pub mod spec;
 pub mod srs;
 pub mod state;
 pub mod stream_probe;
+pub mod stream_statistics;
 pub mod teamspeak;
 pub mod types;